Change format of duration return types to int64 to match Vault

+++ schema/model_app_role_read_token_ttl_response.go
@@ -8,5 +8,5 @@ package schema
 // AppRoleReadTokenTtlResponse struct for AppRoleReadTokenTtlResponse
 type AppRoleReadTokenTtlResponse struct {
 	// The initial ttl of the token to generate
-	TokenTtl int32 `json:"token_ttl,omitempty"`
+	TokenTtl int64 `json:"token_ttl,omitempty"`
 }
